core func was eatracted into a separate project
