+++ core/src/main/java/com/mogikanensoftware/rss2twitter/core/twitter/TwitterPublisher.java
@@ -1,4 +1,4 @@
-package com.mogikanensoftware.rss2twitter.service.twitter;
+package com.mogikanensoftware.rss2twitter.core.twitter;
 
 public interface TwitterPublisher {
